+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-## [Unreleased](https://github.com/dotenvx/dotenvx/compare/v0.10.0...main)
+## [Unreleased](https://github.com/dotenvx/dotenvx/compare/v0.10.1...main)
+
+## 0.10.1
+
+### Changed
+
+* Adjust `precommit` verbosity and coloring
+* Add `--install` flag to precommit - installs to `.git/hooks/pre-commit` ([#31](https://github.com/dotenvx/dotenvx/pull/31))
 
 ## 0.10.0